Refactor/Melhora estilos e uso dos scripts js
