+++ front/public/js/fetchData.js
@@ -5,7 +5,7 @@ const fetchData = async (url) => {
       return data;
   } catch (error) {
       console.error('Error fetching data:', error);
-      throw error; 
+      throw error;
   }
 };
